+++ js/util.js
@@ -1,0 +1,15 @@
+const getRandomPositiveFloat = (a, b, digits = 1) => {
+  const min = Math.min(Math.abs(a), Math.abs(b));
+  const max = Math.max(Math.abs(a), Math.abs(b));
+  const result = Math.random() * (max - min) + min;
+  return +result.toFixed(digits);
+};
+
+const getRandomPositiveInteger = (a, b) => {
+  const min = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
+  const max = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
+  const result = Math.random() * (max - min + 1) + min;
+  return Math.floor(result);
+};
+
+export {getRandomPositiveFloat, getRandomPositiveInteger};
